+++ packages/editor/editor-core/src/__tests__/unit/plugins/media/index.tsx
@@ -399,8 +399,7 @@ describe('Media plugin', () => {
 
     const mediaProvider1 = getFreshMediaProvider();
     await pluginState.setMediaProvider(mediaProvider1);
-    const resolvedMediaProvider1 = await mediaProvider1;
-    await resolvedMediaProvider1.uploadContext;
+    await mediaProvider1;
 
     pluginState.pickers.forEach(picker => {
       picker.setUploadParams = jest.fn();
@@ -408,11 +407,45 @@ describe('Media plugin', () => {
 
     const mediaProvider2 = getFreshMediaProvider();
     await pluginState.setMediaProvider(mediaProvider2);
-    const resolvedMediaProvider2 = await mediaProvider2;
-    await resolvedMediaProvider2.uploadContext;
+    await mediaProvider2;
 
     pluginState.pickers.forEach(picker => {
       expect(picker.setUploadParams as any).toHaveBeenCalledTimes(1);
+    });
+  });
+
+  it.skip('should trigger analytics events for picking and dropzone', async () => {
+    const { pluginState } = editor(doc(p('{<>}')));
+    const spy = jest.fn();
+    analyticsService.handler = spy as AnalyticsHandler;
+
+    afterEach(() => {
+      analyticsService.handler = null;
+    });
+
+    await mediaProvider;
+    await waitForAllPickersInitialised(pluginState);
+
+    const testFileData = {
+      file: {
+        id: 'test',
+        name: 'test.png',
+        size: 1,
+        type: 'file/test',
+      },
+      preview: {
+        dimensions: {
+          height: 200,
+          width: 200,
+        },
+      },
+    };
+
+    (pluginState as any).dropzonePicker!.handleUploadPreviewUpdate(
+      testFileData,
+    );
+    expect(spy).toHaveBeenCalledWith('atlassian.editor.media.file.dropzone', {
+      fileMimeType: 'file/test',
     });
   });
 
@@ -425,9 +458,7 @@ describe('Media plugin', () => {
       analyticsService.handler = null;
     });
 
-    const provider = await mediaProvider;
-    await provider.uploadContext;
-    await provider.viewContext;
+    await mediaProvider;
     await waitForAllPickersInitialised(pluginState);
 
     const testFileData = {
@@ -766,8 +797,7 @@ describe('Media plugin', () => {
         dropzoneContainer,
       );
 
-      const provider = await mediaProvider;
-      await provider.uploadContext;
+      await mediaProvider;
       // MediaPicker DropZone bind events inside a `whenDomReady`, so we have to wait for the next tick
       await sleep(0);
       expect(getWidgetDom(editorView)).toBeNull();
@@ -791,8 +821,7 @@ describe('Media plugin', () => {
         dropzoneContainer,
       );
 
-      const provider = await mediaProvider;
-      await provider.uploadContext;
+      await mediaProvider;
       // MediaPicker DropZone bind events inside a `whenDomReady`, so we have to wait for the next tick
       await sleep(0);
       expect(getWidgetDom(editorView)).toBeNull();