+++ packages/editor/editor-core/src/plugins/media/picker-facade.ts
@@ -10,7 +10,7 @@ import {
   Popup,
   PopupConfig,
 } from '@atlaskit/media-picker';
-import { Context } from '@atlaskit/media-core';
+import { MediaClientConfig } from '@atlaskit/media-core';
 
 import { ErrorReportingHandler } from '@atlaskit/editor-common';
 
@@ -27,7 +27,7 @@ export type ExtendedComponentConfigs = {
 };
 
 export type PickerFacadeConfig = {
-  context: Context;
+  mediaClientConfig: MediaClientConfig;
   errorReporter: ErrorReportingHandler;
 };
 
@@ -69,7 +69,7 @@ export default class PickerFacade {
       picker = this.picker = this.pickerConfig as CustomMediaPicker;
     } else {
       picker = this.picker = await this.mediaPickerFactoryClass(
-        this.config.context,
+        this.config.mediaClientConfig,
         this.pickerConfig as PopupConfig,
       );
     }
@@ -117,17 +117,6 @@ export default class PickerFacade {
     if (this.picker) {
       this.picker.setUploadParams(params);
     }
-  }
-
-  onClose(cb: () => void): () => void {
-    const { picker } = this;
-    if (isPopup(picker)) {
-      picker.on('closed', cb);
-
-      return () => picker.off('closed', cb);
-    }
-
-    return () => {};
   }
 
   show(): void {